fix: Add protection for mutex initialization
Instead of having a loop in the main used for
destroying mutexes there is no a function that
can do it depending on an input number of mutexes
currently in the array. If the initialization of
said mutexes fails it will now call this function
together with a specicif error code and message.