fix: Put gettimeofday before death_thread
If it is in the beginning then the timecode might be offset. For high
philosopher numbers this might get quite high. If its put under all
threads then this function will need an extra mutex, as it will already
be read by the death_thread technically (eventhough a data race would
probably not be possible because there is a usleep at its beginning)